WRG: Suppress Visual Express 2008 warning

+++ Testing/Utilities/siftfast.cpp
@@ -89,7 +89,7 @@ SiftFastImage ReadPGM(FILE *fp)
     image = CreateImage(height, width);
     for (r = 0; r < height; r++)
         for (c = 0; c < width; c++)
-            image->pixels[r*image->stride+c] = ((float) fgetc(fp)) / 255.0;
+            image->pixels[r*image->stride+c] = static_cast<float>(((float) fgetc(fp)) / 255.0);
 
     //Check if there is another image in this file, as the latest PGM
     // standard allows for multiple images.
